Random Strings with 5 different fonts
This program displays 100 copies of a message.  The color and  position of each message is selected at random.  The font of each message is randomly chosen from among five possible fonts.  The messages are displayed on a white background. There is a button that the user can click to redraw the image using new random values.